pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
exclude = ["ENV*", "tests*"]

[tool.setuptools.package-data]
"tempdd.core.configs" = ["*.yaml"]
"tempdd.core.templates" = ["*.md"]
"tempdd.integrations.claude" = [".claude/commands/*.md"]
"tempdd.integrations.gemini" = [".gemini/commands/*.toml"]
"tempdd.integrations.cursor" = [".cursor/commands/*.md"]
"tempdd.integrations.copilot" = [".github/prompts/*.md"]

[project]
name = "tempdd"
version = "0.1.3"
description = "Template-Driven Development Framework for AI-Augmented Coding"
authors = [{name = "Yi-Chang Chen"}]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    'PyYAML',
    'readchar>=4.0.0',
    'rich>=10.0.0'
]

[project.scripts]
tempdd = "tempdd.cli:main"

[project.urls]
Homepage = "https://github.com/GitYCC/TempDD"
Repository = "https://github.com/GitYCC/TempDD"